DockerFRTriton/requirements.txt
torch>=2.0
onnx>=1.14
tritonclient[http]>=2.41.0
numpy>=1.24
Pillow>=10.0
fastapi>=0.110
pydantic>=2.0
uvicorn[standard]>=0.23
python-multipart>=0.0.6
insightface>=0.7.3
onnxruntime>=1.16.0
opencv-python-headless>=4.8.0
scikit-learn>=1.3.0
